+++ openwrite/__init__.py
@@ -5,10 +5,15 @@ from werkzeug.exceptions import HTTPException
 import os
 from .utils.helpers import generate_nonce, get_ip
 import time
-from .utils.models import Settings
+from .utils.models import Settings, Home
+from sqlalchemy import distinct
+import bcrypt
+import json
+from cryptography.hazmat.primitives.asymmetric import rsa
+from cryptography.hazmat.primitives import serialization
+from datetime import datetime, timezone
 
 start_time = time.time()
-
 
 def create_app(test_config=None):
     if test_config is None:
@@ -19,6 +24,130 @@ def create_app(test_config=None):
         db_type = test_config.get("DB_TYPE", "sqlite")
         db_path = test_config.get("DB_PATH", "data.db")
         load_dotenv(test_config.get("env"), override=True)
+
+    if not os.path.exists(".initialized") or test_config is not None:
+        mode = os.getenv("MODE")
+        domain = os.getenv("DOMAIN")
+        f_abs_path = os.path.abspath(__file__)
+        file_cwd = "/".join(f_abs_path.split("/")[:-1])
+        from .utils.create_db import init_db
+        init_db(db_type, db_path)
+        from .utils.models import User, Blog, Post, Settings, Home
+        from .utils.db import init_engine, SessionLocal
+        init_engine(db_type, db_path)
+        from .utils.db import SessionLocal
+        admin_password = os.urandom(16).hex()
+        hashed = bcrypt.hashpw("openwrite".encode("utf-8"), bcrypt.gensalt())
+        admin_user = User(username="admin", email="", password_hash=hashed.decode("utf-8"), verified=1, admin=1)
+        SessionLocal.add(admin_user)
+        SessionLocal.commit()
+        SessionLocal.close()
+        
+        if mode == "single":
+            key = rsa.generate_private_key(
+                public_exponent=65537,
+                key_size=2048
+            )
+
+            private_pem = key.private_bytes(
+                encoding=serialization.Encoding.PEM,
+                format=serialization.PrivateFormat.PKCS8,
+                encryption_algorithm=serialization.NoEncryption()
+            ).decode()
+
+            public_pem = key.public_key().public_bytes(
+                encoding=serialization.Encoding.PEM,
+                format=serialization.PublicFormat.SubjectPublicKeyInfo
+            ).decode()
+
+            now = datetime.now(timezone.utc).replace(microsecond=0)
+
+            new_blog = Blog(
+                owner=1, 
+                name="default", 
+                title=domain, 
+                index="on", 
+                access="domain",
+                description_raw=f"![hello](https://openwrite.b-cdn.net/hello.jpg =500x258)\n\n# Hello there! 👋\n\nYou can edit your blog description in [dashboard](http://{domain}/dashboard/edit/default)",
+                description_html=f"<p><img src=\"https://openwrite.b-cdn.net/hello.jpg\" width=\"500\" height=\"258\"></p><h1>Hello there! 👋</h1><p>You can edit your blog description in <a href=\"http://{domain}/dashboard/edit/default\">dashboard</a></p>",
+                css="",
+                pub_key=public_pem,
+                priv_key=private_pem,
+                theme="default",
+                created=now
+            )
+            SessionLocal.add(new_blog)
+
+        hometext_pl = """
+                 <h1 class="centered"><span style="color: #f2d2c8">cicha</span> przestrzeń na <span style="color: #5a36bf">głośne</span> myśli.</h1>
+                
+                <article class="main-content" style="text-align: center">
+                    <p><strong>Openwrite</strong> to lekka, otwartoźródłowa platforma blogowa stworzona z myślą o wolności pisania – bez reklam, trackerów i zbędnego balastu.</p><br>
+
+          - Pisz w Markdownie<br>
+          - Zainstaluj w kilka minut<br>
+          - Publikuj do sieci, przez Gemini lub na Mastodonie – jak chcesz<br>
+
+        <br>
+        <p>Bez analityki. Bez ukrytych skryptów.<br>
+        Tylko Ty, Twoje myśli i miejsce, w którym możesz je wyrazić.</p>
+        <p>Wybierz motyw. Dodaj obrazki. Śledź wyświetlenia.<br>
+        Zbuduj bloga po swojemu – minimalistycznego lub z dodatkami.</p>
+        <p><strong>To Twoje słowa. Twój blog. Twoje zasady.</strong><br><br>
+        Zacznij pisać — i odzyskaj głos.</p>
+                    </p>
+                </article>
+                <div class="action-links">
+                    <a class="btn purple" href="/register">Zarejestruj się tutaj</a>
+                    <span class="or-separator">lub</span>
+                    <a class="btn empty" href="https://main.openwrite.io/hosting">Samodzielny hosting</a>
+                </div>
+        """
+
+        hometext_en = """
+                 <h1 class="centered"><span style="color: #f2d2c8">quiet</span> space for <span style="color: #5a36bf">loud</span> thoughts.</h1>
+                
+                <article class="main-content" style="text-align: center">
+                    <p><strong>Openwrite</strong> is a lightweight, open-source blogging platform built for people who just want to write — without ads, trackers, or bloated nonsense.</p><br>
+
+          - Write in Markdown<br>
+          - Self-host in minutes<br>
+          - Publish to the open web, Gemini, or Mastodon — your choice<br>
+
+        <br>
+        <p>There’s no analytics. No hidden scripts. No engagement traps.<br>
+        Just you, your thoughts, and an honest place to share them.</p>
+        <p>Choose a theme. Add images. Track views.<br>
+        Build your blog your way — minimal or feature-rich.</p>
+        <p><strong>It’s your voice. Your blog. Your rules.</strong><br><br>
+        Start writing — and own your words.</p>
+                </article>
+                <div class="action-links">
+                    <a class="btn purple" href="/register">Register Here</a>
+                    <span class="or-separator">or</span>
+                    <a class="btn empty" href="https://main.openwrite.io/hosting">Self Host</a>
+                </div>
+        """
+
+        with open(f"{file_cwd}/utils/i18n.json", "r", encoding='utf-8') as f:
+            data = json.load(f)
+            
+        for lang, entries in data.items():
+            for key, value in entries.items():
+                SessionLocal.add(Home(language=lang, name=key, type="translation", content=value))
+
+        new_setting = Settings(name="logo", value="/static/logo.png")
+        new_home_pl = Home(language="pl", name="hometext", type="text", content=hometext_pl)
+        new_home_en = Home(language="en", name="hometext", type="text", content=hometext_en)
+        SessionLocal.add(new_setting)
+        SessionLocal.add(new_home_pl)
+        SessionLocal.add(new_home_en)
+        SessionLocal.commit()
+
+        if test_config is None:
+            with open(".initialized", "w") as f:
+                f.write("ok")
+
     app = Flask(__name__, template_folder="templates", subdomain_matching=True, static_url_path='/static')
     app.wsgi_app = ProxyFix(app.wsgi_app, x_for=1, x_proto=1, x_host=1, x_port=1)
     app.secret_key = os.getenv("SECRET_KEY")
@@ -31,11 +160,9 @@ def create_app(test_config=None):
     from .routes.main import main_bp
     from .routes.upload import upload_bp
     from .routes.federation import federation_bp
-    from .utils.translations import load_translations
     from .utils.db import init_engine, SessionLocal
     from flask_cors import CORS
     
-    translations = load_translations()
     CORS(app)
 
     init_engine(db_type, db_path)    
@@ -49,9 +176,18 @@ def create_app(test_config=None):
     app.register_blueprint(upload_bp)
     app.register_blueprint(federation_bp)
 
+    def _(key: str):
+        return g.trans.get(key, key)
+
+    app.jinja_env.globals['_'] = _
+
     @app.before_request
     def before():
         from flask import request, session
+        from openwrite.utils.models import Home, Settings
+        g.db = SessionLocal
+        t_query = g.db.query(distinct(Home.language)).all()
+        translations = [row[0] for row in t_query]
         lang = request.cookies.get("lang")
         if not lang or lang not in translations:
             accept = request.headers.get("Accept-Language", "")
@@ -63,13 +199,14 @@ def create_app(test_config=None):
         if lang not in translations:
             lang = "en"
 
+        trans = g.db.query(Home).filter(Home.type == "translation", Home.language == lang).all()
+        
         g.mode = os.getenv("MODE")
         f_abs_path = os.path.abspath(__file__)
         g.mainpath = "/".join(f_abs_path.split("/")[:-1])
-        g.trans = translations[lang]
+        g.trans = {t.name: t.content for t in trans if len(t.content) > 1}
         g.alltrans = translations
         g.lang = lang
-        g.db = SessionLocal
         g.main_domain = os.getenv("DOMAIN")
         g.blog_limit = os.getenv("BLOG_LIMIT")
         g.register_enabled = os.getenv("SELF_REGISTER", "no") == "yes"
@@ -91,12 +228,14 @@ def create_app(test_config=None):
 
     @app.context_processor
     def inject_globals():
+        from openwrite.utils.models import Home
+        langs = g.db.query(Home).filter_by(name="lang_name").all()
+        lang_obj = {}
+        for l in langs:
+            lang_obj[l.language] = {"name": l.content}
         return {
             'current_lang': g.lang,
-            'available_languages': {
-                'en': {'name': 'English', 'flag': '🇬🇧'},
-                'pl': {'name': 'Polski', 'flag': '🇵🇱'}
-            }
+            'available_languages': lang_obj
         }
 
     @app.after_request